backend/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
scapy==2.5.0
pyroute2==0.7.12
